requirements.txt
piexif
pillow
black
flake8
pydocstring
pycodestyle
